04-model-gtw/requirements.txt
--prefer-binary
azure-ai-projects>=2.0.0
azure-identity>=1.21.0
requests>=2.32.0
